feat(ai): 提取共用的 ai 输入框组件
